fixed parse error to calculate decimals correctly

+++ reactcalc/src/App.js
@@ -51,26 +51,26 @@ class App extends Component {
     if (this.state.operator == "plus") {
       this.setState({
         input:
-          parseInt(this.state.previousNumber) +
-          parseInt(this.state.currentNumber),
+          parseFloat(this.state.previousNumber) +
+          parseFloat(this.state.currentNumber),
       });
     } else if (this.state.operator == "subtract") {
       this.setState({
         input:
-          parseInt(this.state.previousNumber) -
-          parseInt(this.state.currentNumber),
+          parseFloat(this.state.previousNumber) -
+          parseFloat(this.state.currentNumber),
       });
     } else if (this.state.operator == "multiply") {
       this.setState({
         input:
-          parseInt(this.state.previousNumber) *
-          parseInt(this.state.currentNumber),
+          parseFloat(this.state.previousNumber) *
+          parseFloat(this.state.currentNumber),
       });
     } else if (this.state.operator == "divide") {
       this.setState({
         input:
-          parseInt(this.state.previousNumber) /
-          parseInt(this.state.currentNumber),
+          parseFloat(this.state.previousNumber) /
+          parseFloat(this.state.currentNumber),
       });
     }
   };
